requirements/base.txt
pytz==2018.4  # https://github.com/stub42/pytz
Pillow==5.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==18.1.0  # https://github.com/hynek/argon2_cffi
redis>=2.10.5  # https://github.com/antirez/redis

# Django
# ------------------------------------------------------------------------------
django==2.0.4  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.4  # https://github.com/joke2k/django-environ
django-model-utils==3.1.1  # https://github.com/jazzband/django-model-utils
django-allauth==0.35.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-redis==4.9.0  # https://github.com/niwinz/django-redis

# Django REST Framework
djangorestframework==3.8.2  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client

#graphQL
graphene==2.0.1
graphene-django==2.0.0
django-filter==1.1.0
django-graphql-jwt==0.1.5

# Django CORS
django-cors-headers==2.2.0